requirements.txt
Django==4.2
djangorestframework
mysqlclient
django-cors-headers
